solve 2357 - 최솟값과 최댓값
gold 1
값이 바뀌는 쿼리는 들어오지 않지만 최솟값과 최댓값을 보관하는 트리 두 개를 동시에 만들어야 한다.

+++ Algorithm/Algorithm/2042 - 구간 합 구하기.cpp
@@ -60,7 +60,7 @@ int main()
 	}
 	init(arr, segtree, 1, 0, n - 1);
 
-	int query, a, b;
+	long long query, a, b;
 	for (int i = 0; i < m + k; ++i) {
 		cin >> query >> a >> b;
 		if (query == 1) {
